move Client to ClientFragment
